+++ src/lib/schema-tag.ts
@@ -1,8 +1,6 @@
 import z, { core } from "zod";
 import { createRenderable } from "./core/renderable";
-import { interpolate } from "./core/interpolate";
 import { TypedTag } from "./typed-tag";
-import { InterpolationError } from "../main";
 
 export const schemaTag = <
     S extends core.$ZodShape,
@@ -16,19 +14,7 @@ export const schemaTag = <
         strs: TemplateStringsArray,
         ...vals: any[]
     ) {
-        return createRenderable(
-            function renderRenderable(karg) {
-                const parsedkarg = schema.safeDecode(karg)
-                if (parsedkarg.error) throw InterpolationError.for(parsedkarg.error, {
-                    renderer: this,
-                    index: -1,
-                    op: 'root-schema',
-                    strings: [...strs],
-                    value: schema,
-                })
-                return interpolate.call(this, parsedkarg.data,  strs, ...vals)
-            }
-        )
+        return createRenderable(strs, vals, schema)
     }
 
     return schemaTag as TypedTag<