+++ src/lib/zt/identity.ts
@@ -1,0 +1,3 @@
+import { createRenderable } from "../core/renderable";
+
+export const tagIdentity = createRenderable<void, []>([''], [])