+++ src/lib/zt/if-condition.ts
@@ -1,6 +1,6 @@
 import { type IRenderable } from "../core/renderable";
-import { typedTag } from "../typed-tag";
+import { tagIdentity } from "./identity";
 
 export const ifCondition = <T extends IRenderable<any, any>>(condition: unknown, template: T) => {
-    return !!condition ? template : typedTag``
+    return !!condition ? template : tagIdentity
 }